Add some failure tests for "closeLong"

+++ contracts/Hyperdrive.sol
@@ -413,7 +413,7 @@ abstract contract Hyperdrive is MultiToken {
     /// @notice Closes a long position with a specified maturity time.
     /// @param _maturityTime The maturity time of the short.
     /// @param _bondAmount The amount of longs to close.
-    function closeLong(uint32 _maturityTime, uint256 _bondAmount) external {
+    function closeLong(uint256 _maturityTime, uint256 _bondAmount) external {
         if (_bondAmount == 0) {
             revert Errors.ZeroAmount();
         }
@@ -544,7 +544,7 @@ abstract contract Hyperdrive is MultiToken {
     /// @notice Closes a short position with a specified maturity time.
     /// @param _maturityTime The maturity time of the short.
     /// @param _bondAmount The amount of shorts to close.
-    function closeShort(uint32 _maturityTime, uint256 _bondAmount) external {
+    function closeShort(uint256 _maturityTime, uint256 _bondAmount) external {
         if (_bondAmount == 0) {
             revert Errors.ZeroAmount();
         }
